Add primitive database layer
