+++ internal/database/database.go
@@ -7,6 +7,25 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 )
 
+type Operator struct {
+	id        int64
+	UserName  string
+	FirstName string
+	LastName  string
+}
+
+type DBModel interface {
+	Insert()
+	Delete()
+}
+
+func (op *Operator) Insert() {
+	op.id = createOperator(op.UserName, op.FirstName, op.LastName)
+}
+
+func (op *Operator) Delete() {
+	deleteOperator(op.id)
+}
 
 func CreateSchema() {
 	db, err := sql.Open("sqlite3", "duty.db")
